requirements.txt
# FastAPI and async
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.1

# HTTP and requests
httpx>=0.24.0
requests>=2.31.0

# System utilities
psutil>=5.9.5
python-dotenv>=1.0.0

# Git operations
gitpython>=3.1.32

# Memory systems (optional - install separately if needed)
# chromadb
# neo4j

# Machine Learning
scikit-learn>=1.3.0

# LLM / GGUF Support (LOCAL OFFLINE INFERENCE)
llama-cpp-python>=0.2.50
huggingface-hub>=0.20.0

# GitHub API
PyGithub>=2.1.1

# Pydantic
pydantic>=2.4.0
pydantic-settings>=2.1.0

# Testing
pytest>=7.4.0
